[IT-427] Fix rotate credentials lambda (#75)
The lambda broke because 'enable password expiration' was disabled
in synapsedev and synapseprod accounts.  The lambda uses the
PasswordPolicy.MaxPasswordAge as the expiration age.  Since it was
disabled the API response did not contain that key.  We now de-couple
the password expiration age from the access key expiration age.
The expiration period for password and access key are now independent
and the key expiration age is passed in from the template that
deploys the rotate credentials lambda.