Setup cloudformation tests with taskcat
Use taskcat[1] to test cloudformation templates.  Taskcat tests
templates by deploying resources to an AWS account then tearing
them all down afterwards.  This makes sure that templates can actually
provision resources to the AWS account in a specified AZ.  We
only test on "us-east-1" because that's the only AZ we use.

Taskcat is a bit opinionated about it's directory structure so
we could not include the rotate-credentials lambda.  We also
have a few issues with CF templates that make them difficult to test:

https://sagebionetworks.jira.com/browse/IT-475
https://sagebionetworks.jira.com/browse/IT-476

[1] https://github.com/aws-quickstart/taskcat