[IT-497] make SynapseUserName an optional parameter (#93)
Users may only want to provision a generic bucket.  The SynapseUserName
parameter is not applicable in those cases.  Do not not deploy a
owner.txt file in those cases.

This requires refactoring the sceptre "synapse_bucket_notify" hook[1] to
not create an owner.txt file when `synapse_username` is not specified.

[1] https://github.com/Sage-Bionetworks/sceptre-provisioner-hooks/pull/4