Set return code to 0 when deploying windows instances
The return code does not seem to work consistently causing the
WindowsServerWaitHandle to time.  This results in a failed
provision of windows instances.  Make it always return 0 for no
error.

The drawback here is that we may not notice a failed deployment.
I'm hoping this will be rare and we will have this issue fixed
shortly.  When fixed we should put back the `ERRORLEVEL%`