[IT-553] Restrictive access to EC2 instances (#133)
Instead of giving a group access to a provisioned EC2 we change to
only give the EC2 owner access to the provisioned instance.  Jumpcloud
configurations for AWS linux and ubuntu requests are handled in the
CFN template. Jumpcloud config for windows is handled in a powershell
script updated in PR https://github.com/Sage-Bionetworks/infra-utils/pull/33