Template for Gateway VPC endpoints
Add a template to provision a Gateway VPC endpoint[1]. What an endpoint
effectively does is enable instances in your VPC to use their private
IP addresses to communicate with resources in other services.
Consequently, your instances do not require public IP addresses, and
you do not need an Internet Gateway, a NAT instance, or a virtual
private gateway in your VPC. This is because you use endpoint policies
to control access to resources in other services. Traffic between your
VPC and the AWS service does not leave the Amazon network which will
reduce latency and cost.

[1] https://docs.aws.amazon.com/vpc/latest/userguide/vpce-gateway.html