[IT-831] Multiple synapse user and teams (#251)
Update template to support multiple user and teams.  Username in
owner file is deprecated so we change to using synapse user and
team IDs.

Need to change to jinja template because synapse owner file
only supports synapse IDs on multiple lines.  The easiest way
to get the IDs on multiplel was to use a jinja transform.

The synapse IDs should be passed in like so..
```
sceptre_user_data:
  SynapseIDs: ["1300139", "3390915"]
```