Pin pre-commit hooks
cfn-lint v0.34.0 is breaking build with the following warnings:

"W4002 As the resource "metadata" section contains reference to a "NoEcho"
parameter JcSystemsGroupId, CloudFormation will display the parameter
value in plaintext.."

This will  pin cfn-lint hook to v0.33.2 so we don't get those silly warnings.
We also pin the other pre-commit hooks to the curret latest version.