Add service user template (#299)
In this PR, I'm adding a template for "service users", which was derived from the `IAM/service-account.yaml` template. The only difference is that I removed the IAM role resource. The use case for this template is for tools that don't support IAM roles (_e.g._ Nextflow). 

Once this PR is merged, I'll need a new tag to be minted because I plan on using this right away in [aws-workflows-nextflow-infra](https://github.com/Sage-Bionetworks-Workflows/aws-workflows-nextflow-infra).