[SC-373] Separate cost center tags into two lists
The service catalog's “number of values per tag” hard limit of 25
values means that we cannot provide more than 25 values in a tag option.
To workaround this problem we will provide a tag option with less than
25 program codes (SageFinancialProgramCodes.json).  If it's not
available in the tag option then the user will need to create a custom
tag with the other program codes (SageFinancialProgramCodesOther.json).