fork PyPlate for pining
We use PyPlate in the service catalog and currently we are getting it
from the upstream master branch[1].  The problem with that dependency
is that it's not tagged upstream so we cannot pin to a specific version.
Adding it to our repo allows us to tag and pin to a specific version
which will make it less prone to error.

The code is taken from awslabs as-is.

[1] https://github.com/awslabs/aws-cloudformation-templates/tree/master/aws/services/CloudFormation/MacrosExamples/PyPlate