[IT-1234] AWS Security group for TGW
We use the transit gateway to route traffic between the hub
and spoke VPCs.  To allow access to resources in a spoke VPC we need
to give the hub VPC access to the resources.  This adds an SG ingress
to resources provisioned in our spoke VPCs which will allow VPN users
access to the resources in the spoke VPCs.