[IT-2052] Replace `jq` filter with api filter
There was a recent failure while deploying the 'jharker-test' stack in
the 'sandbox-provisioner' project during the '02_associate_jc_systems'
command. The resulting virtual machine did not have the required
association and was not accessible using ssh keys from jumpcloud.

The underlying user ID could not be found because it does not appear in the
default list of system users, but can be found when searching based on
email address.

Replace the client-side `jq` filter with a server-side api filter.