[IT-2075] Allow bucket collaboration (#351)
Occasionally we want to allow our external collaborators
to use the AWS CLI to upload files to buckets that we provision.
The bucket template to create an external synapse bucket does
not support that use case currently.  We need to update the
permissions to allow for that use case. Change from BucketOwnerPreferred
to BucketOwnerEnfored to disable ACLs and set the bucket owner
to autmatically have full control over every object in the bucket.