[IT-2268] Add support for CustomerManagedPolicyReference
In order to use a custom managed policy for granting access to the cost
explorer, we need to add support for at least one customer-managed
policy to the managed-policy template. Adding support for multiple
customer-managed policies in a single permission set will be more
difficult due to the data structure of the parameter.

Updates to cfn-lint are required to avoid an E3002, and the updates also
introduced new E2531 and E1001 occurrences - the former is a deprecation
error for python 3.6, and the latter is related to org-formation syntax.