[IT-1665] Remove travis service user (#363)
This template is old and we shouold not be using the travis
user that's created from it anymore.  We should not be using
CI users that are created from org formation access deployments[1]

NOTE: This may cause build failures in travis-ci because really
old CI configs might still be using these service users.  The
strategy is to fix up as needed.

[1] https://github.com/Sage-Bionetworks-IT/organizations-infra/tree/master/org-formation/600-access