[IT-2768] Remove importvalue from essentials template
Using an `!ImportValue` makes one template depend on another.
In this case the essentials stack depends on the OIDC stack.
When we attempted to update the OIDC stack in PR https://github.com/Sage-Bionetworks-IT/organizations-infra/pull/859
we got the following error..

```
Export us-east-1-sagebase-github-oidc-sage-bionetworks-it-ProviderRoleArn-organizationsinfra
cannot be deleted as it is in use by essentials
```

This required a temporary fix with PR https://github.com/Sage-Bionetworks-IT/organizations-infra/pull/863

The right fix for this problem is to make the dependency a little
looser.  We setup a parameter to pass in the values for the dependency.